+++ zython/_compile/zinc/to_str.py
@@ -213,6 +213,7 @@ class Op2StrType(UserDict):
         self[_Op_code.decreasing] = partial(_global_constraint, "decreasing")
         self[_Op_code.strictly_decreasing] = partial(_global_constraint, "strictly_decreasing")
         self[_Op_code.cumulative] = partial(_global_constraint, "cumulative")
+        self[_Op_code.disjunctive] = partial(_global_constraint, "disjunctive")
         self[_Op_code.table] = partial(_global_constraint, "table", flatten_args=False)
 
     def __missing__(self, key):  # pragma: no cover